requirements.txt
wetro==0.1.14
python-decouple==3.8